% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tidy_2d.R
\name{plot_tidy_2d}
\alias{plot_tidy_2d}
\title{plot_tidy_2d}
\usage{
plot_tidy_2d(nn, X)
}
\arguments{
\item{nn}{the neural network.}

\item{X}{the X data. A data.frame that is 2xN}
}
\value{
a ggplot
}
\description{
plot a 2d scatter plot with confidence tiles.
}
