% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential.R
\name{Sequential}
\alias{Sequential}
\title{Sequential neural networks}
\description{
Setup a sequential (uni-directional) deep
feed forward neural network. The Sequential is used as a
wrapper for the modules.
}
\note{
Note that delta can refer to dLdA or dLdZ over the
course of the for loop, depending on the module m
}
\seealso{
Other architecture: 
\code{\link{Linear}}
}
\concept{architecture}
\section{Super class}{
\code{\link[neuralnetr:ClassModule]{neuralnetr::ClassModule}} -> \code{Sequential}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{loss}}{The loss module}

\item{\code{modules}}{the list of current modules}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sequential$new()}}
\item \href{#method-sgd}{\code{Sequential$sgd()}}
\item \href{#method-forward}{\code{Sequential$forward()}}
\item \href{#method-backward}{\code{Sequential$backward()}}
\item \href{#method-sgd_step}{\code{Sequential$sgd_step()}}
\item \href{#method-classify}{\code{Sequential$classify()}}
\item \href{#method-print_accuarcy}{\code{Sequential$print_accuarcy()}}
\item \href{#method-clone}{\code{Sequential$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
setup the neural network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$new(modules, loss)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modules}}{list of modules}

\item{\code{loss}}{the loss module}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sgd"></a>}}
\subsection{Method \code{sgd()}}{
train neural network using stochastic gradient descent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$sgd(X, Y, iters = 100, lrate = 0.005)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{the X input (m x b)}

\item{\code{Y}}{the Y (target) input}

\item{\code{iters}}{amount of iterations.}

\item{\code{lrate}}{the learning rate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forward"></a>}}
\subsection{Method \code{forward()}}{
Compute Ypred
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$forward(Xt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt}}{the input at time t}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-backward"></a>}}
\subsection{Method \code{backward()}}{
Update dLdW and dLdW0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$backward(delta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{the backpropagated error}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sgd_step"></a>}}
\subsection{Method \code{sgd_step()}}{
gradient descent step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$sgd_step(lrate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lrate}}{the learning rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classify"></a>}}
\subsection{Method \code{classify()}}{
classify the labels of the input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$classify(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{input X}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_accuarcy"></a>}}
\subsection{Method \code{print_accuarcy()}}{
print accuracy during training
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$print_accuarcy(it, X, Y, cur_loss, every = 250)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{it}}{iteration number}

\item{\code{X}}{data X}

\item{\code{Y}}{target Y}

\item{\code{cur_loss}}{the current loss}

\item{\code{every}}{how often should the function return feeddback?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
