% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigmoid.R
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid activation}
\description{
Each activation module has a forward method that
takes in a batch of pre-activations Z and returns a batch of
activations A.

Each activation module has a backward method that takes in dLdA and
returns dLdZ, with the exception of SoftMax, where we assume dLdZ is
passed in.
}
\seealso{
Other activation: 
\code{\link{ReLU}},
\code{\link{SoftMax}}
}
\concept{activation}
\section{Super class}{
\code{\link[neuralnetr:ClassModule]{neuralnetr::ClassModule}} -> \code{Sigmoid}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{A}}{the activation vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-forward}{\code{Sigmoid$forward()}}
\item \href{#method-backward}{\code{Sigmoid$backward()}}
\item \href{#method-clone}{\code{Sigmoid$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="neuralnetr" data-topic="ClassModule" data-id="sgd_step">}\href{../../neuralnetr/html/ClassModule.html#method-sgd_step}{\code{neuralnetr::ClassModule$sgd_step()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forward"></a>}}
\if{latex}{\out{\hypertarget{method-forward}{}}}
\subsection{Method \code{forward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sigmoid$forward(Z)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Z}}{a vector of pre-activations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector of activations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-backward"></a>}}
\if{latex}{\out{\hypertarget{method-backward}{}}}
\subsection{Method \code{backward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sigmoid$backward(dLdA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dLdA}}{vector of gradients.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a vector gradeints.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sigmoid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
