% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftMax.R
\name{SoftMax}
\alias{SoftMax}
\title{SoftMax}
\description{
Classic softmax that can be interpreted as a probability
distribution over n items

For `SoftMax$class_fun()`, given the column
vector of class probabilities for each point (computed by Softmax), return
a vector of the classes (integers) with the highest probability for each point.
}
\seealso{
Other activation: 
\code{\link{ReLU}},
\code{\link{Sigmoid}}
}
\concept{activation}
\section{Super class}{
\code{\link[neuralnetr:ClassModule]{neuralnetr::ClassModule}} -> \code{SoftMax}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-forward}{\code{SoftMax$forward()}}
\item \href{#method-backward}{\code{SoftMax$backward()}}
\item \href{#method-class_fun}{\code{SoftMax$class_fun()}}
\item \href{#method-clone}{\code{SoftMax$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="neuralnetr" data-topic="ClassModule" data-id="sgd_step">}\href{../../neuralnetr/html/ClassModule.html#method-sgd_step}{\code{neuralnetr::ClassModule$sgd_step()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forward"></a>}}
\if{latex}{\out{\hypertarget{method-forward}{}}}
\subsection{Method \code{forward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftMax$forward(Z)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Z}}{a vector of pre-activations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
probabilities of the given class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-backward"></a>}}
\if{latex}{\out{\hypertarget{method-backward}{}}}
\subsection{Method \code{backward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftMax$backward(dLdZ)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dLdZ}}{a vector of gradients dLdZ.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dLdZ
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-class_fun"></a>}}
\if{latex}{\out{\hypertarget{method-class_fun}{}}}
\subsection{Method \code{class_fun()}}{
a classify function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftMax$class_fun(Ypred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Ypred}}{a vector of predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the index for the vector of the
predicted label.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoftMax$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
