% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xor_example}
\alias{xor_example}
\title{XOR problem}
\format{list with matrix X (coordinates in 2D space) and matrix Y (binary labels)}
\usage{
xor_example
}
\description{
Classic XOR problem
}
\concept{datasets}
\keyword{datasets}
