% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_2d_grid.R
\name{classify_2d_grid}
\alias{classify_2d_grid}
\title{Classify 2d plot}
\usage{
classify_2d_grid(nn, xlims = c(0, 2), ylims = c(0, 2), bys = 0.1)
}
\arguments{
\item{nn}{the neural network.}

\item{xlims}{vector. xlimits.}

\item{ylims}{vector. ylimits.}

\item{bys}{how fine should the grid be?}
}
\value{
data.frame
}
\description{
convert a scatter plot into a grid-like plot of preditction confidence.
}
