% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLL.R
\name{NLL}
\alias{NLL}
\title{Negative Log Liklihood}
\description{
Each loss module has a forward method that takes in a batch of predictions Ypred
(from the previous layer) and labels Y and returns a scalar loss value.

The NLL module has a backward method that returns dLdZ, the gradient with respect
to the preactivation to SoftMax (note: not the activation!), since we are always
pairing SoftMax activation with NLL loss
}
\section{Super class}{
\code{\link[neuralnetr:ClassModule]{neuralnetr::ClassModule}} -> \code{Module}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-forward}{\code{NLL$forward()}}
\item \href{#method-backward}{\code{NLL$backward()}}
\item \href{#method-clone}{\code{NLL$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="neuralnetr" data-topic="ClassModule" data-id="sgd_step">}\href{../../neuralnetr/html/ClassModule.html#method-sgd_step}{\code{neuralnetr::ClassModule$sgd_step()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-forward"></a>}}
\subsection{Method \code{forward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NLL$forward(Ypred, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-backward"></a>}}
\subsection{Method \code{backward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NLL$backward()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NLL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
